Removed the dependency of three Compartments and extended it to N possible compartments and refactored the code.
Population2.csv: Instead of the number of people and the initial infected, the csv now holds the initial values for each compartment. They need to be listed in the same order as in the Placeholders.csv and SIR.csv (or the compartment_adjacency.csv). Also, the x and y positions need to be and the end in the order x,y. If these conditions are not met, the script will not be able to calculate everything right.